% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actives.R
\name{actives}
\alias{actives}
\title{Determine when individual is active or not}
\usage{
actives(df, minSpeed = NULL)
}
\arguments{
\item{df}{A data frame containing x, y coordinates and speed columns named "x.pos", "y.pos", "speed" for a fragment}

\item{minSpeed}{A numeric value expressed in the same unit than speed, corresponding to the treschold above which individual
is considered as active}
}
\value{
this function returns a vector containing True or False when individual is active or not respectively
}
\description{
Given a data frames containing tracking informations for a given fragment including speed,
this function return a vector containing True or False when individual is active or not respectively
}
\examples{

#TODO

}
