% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseBouts.R
\name{analyseBouts}
\alias{analyseBouts}
\title{Bouts analyses}
\usage{
analyseBouts(trackDat, Bstate = NA, Time = NA, pattern = NA)
}
\arguments{
\item{trackDat}{A list of data frame containing tracking informations for each fragment}

\item{Bstate}{The name of a vector/df column present in each df of trackDat and indicating the state of the individual
along a fragment}

\item{Time}{The name of a vector/df column present in each df of trackDat and indicating the time
of each Bstate along the fragment (optional)}

\item{pattern}{A character string containing a regular expression (or character string for fixed = TRUE)
to be matched in the Bstate vector. see gregexpr()}
}
\value{
this function returns a new list of dataframe corresponding to each new fragments with the result of customFunc appended
}
\description{
Given a list of data frames containing tracking informations and a vector
containing behavioral patterns (e.g., behavioral states, location in areas) for each fragment,
the function will look for the specified pattern and compute some simple metrics about it
}
\examples{

#TODO

}
