% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseTime.R
\name{analyseTime}
\alias{analyseTime}
\title{perform analysis across fragments and time}
\usage{
analyseTime(
  trackDat,
  timeCol = NULL,
  customFunc = NULL,
  Tinterval = NULL,
  Tstep = 1,
  sampling = 1,
  wtd = FALSE
)
}
\arguments{
\item{trackDat}{A list of data frame containing tracking informations for each fragment (including a timeline)}

\item{timeCol}{A character string specifying the name of the timeline column}

\item{customFunc}{A function used to perform the computation across time}

\item{Tinterval}{A vector containing two numeric values expressed in the timeline unit and
specifying the time interval on which the computation is performed
(default is null, meaning the computation will be performed on the whole timeline)}

\item{Tstep}{A numeric value expressed in the timeline unit and specifying the size of the
sliding window used to perform the computation}

\item{sampling}{A numeric value expressed in the timeline unit and specifying a subsampling used to
to perform the computation, allow to make computation faster, it hence determine the resolution of the
returned results (e.g., 5000 mean that values will be computed every 5000 frames)}

\item{wtd}{TRUE or FALSE, compute a weighed metric (TRUE) or not (FALSE), (default is FALSE)}
}
\value{
this function returns a vector containing the smoothed values
computed according to the custom function across time
}
\description{
Given a list of data frames containing tracking informations for each fragment (including the timeline)
and a custom function, this function perform the computation specified by the custom function across time
and smooth it before returning values
}
\examples{

#TODO

}
