% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANND.R
\name{ANND}
\alias{ANND}
\title{Compute Average Nearest Neighbour Distance among fragments}
\usage{
ANND(
  trackDat,
  timeCol = NULL,
  Tinterval = NULL,
  sampling = NULL,
  scale = NULL,
  bootn = 0
)
}
\arguments{
\item{trackDat}{A list of data frame containing tracking informations for each fragment (including a timeline)}

\item{timeCol}{A character string specifying the name of the timeline column}

\item{Tinterval}{A vector containing two numeric values expressed in the timeline unit and
specifying the time interval on which the computation is performed
(default is null, meaning the computation will be performed on the whole timeline)}

\item{sampling}{A numeric value expressed in the timeline unit and specifying a subsampling used to
to perform the computation, allow to make computation faster, it hence determine the resolution of the
returned results (e.g., 5000 mean that values will be computed every 5000 frames)}

\item{bootn}{A numeric value corresponding to the number of bootstrap sampling to compute studentize 95\% CI}
}
\value{
this function returns a list containing three sublist, the first sublist contains a dataframe with
bootstrap results (mean ANND, CI 97.5\%, CI 2.5\%, Sd, number of detected individual
and the resampled timeline according to timeCol argument),
the second sublist contains raw neighbour distances computed across each time unit.
The third list returns two sublists, the first one containing, containing dataframes
with the fragments sampled over the bootstrap iterations, while the second contains the values of the nearest
neighbour distance for each of the sampled fragments.
}
\description{
Given a list of data frames containing tracking informations for each fragment (including the timeline)
this function returns the Average Nearest Neighbour Distance among fragments along a specified timeline. Also,
if bootn parameter is above 0, the function compute studentize 95 \% CI using bootstrapping method
}
\examples{

#TODO

}
