% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3ConvertFunc.R
\name{as.data.frame.varList}
\alias{as.data.frame.varList}
\title{Converting varList or tracklets objects using S3 method.}
\usage{
\method{as.data.frame}{varList}(x, row.names = NULL, optional = FALSE, ...)
}
\value{
these functions return the converted varList or tracklets objects.
}
\description{
as.data.frame method to convert varList to a dataframe.
}
\author{
Quentin PETITJEAN
}
\keyword{internal}
