% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.R
\name{circles}
\alias{circles}
\title{Generate points located on circles contour and draw circle(s).}
\usage{
circles(
  x = NULL,
  y = NULL,
  radius = NULL,
  Res = 500,
  center = NULL,
  col = NULL,
  border = "black",
  lwd = 1,
  lty = 1,
  draw = T
)
}
\arguments{
\item{x}{A coordinate vector of circles to plot.}

\item{y}{A coordinate vector of circles to plot.}

\item{radius}{A single numeric value or a vector specifying the radius of the circles to plot.}

\item{Res}{A numeric value corresponding to the resolution of the circles (i.e., the number of points used to draw the circles contour, default = 500).}

\item{center}{Either TRUE or a single integers or a vector of integers specifying the symbol(s) or a single character to be used to represent the circle center
(see points for possible values and their interpretation).}

\item{col}{The color or a vector of colors for filling the circles, the default leave polygons unfilled.}

\item{border}{The color or a vector of colors to draw the border of the circles (default = "black").}

\item{lwd}{The value of line width or a vector of line width of the circles border (default = 1).}

\item{lty}{An integer or a vector of integer corresponding to line type to be used for drawing circles border, as in par (default = 1).}

\item{draw}{A logical value (i.e., TRUE or FALSE) indicating whether the circle(s) should be drawn or not (default = TRUE).}
}
\value{
Draw the circles from center coordinates and value of the radius on an existing plot window or, if there is no active plot window, on a new plot window.
It also returns the coordinates of the points simulated to draw the contour of the circle(s).
}
\description{
Given x and y coordinates as well as the circle radius, this function draw circle(s) and returns the coordinates of the points simulated to draw the contour of the circle.
}
\examples{

# draw 10 red-border and red-filled circles of different size on a new plot
circles(
 x = sample(1:100, 10),
 y = sample(1:100, 10),
 radius = sample(1:10, 10),
 center = TRUE,
 border = "red",
 col = adjustcolor("firebrick", alpha = 0.2),
 Res = 500,
 lwd = 1.5,
 lty = 1
)

# draw 5 red-border and red-filled circles and 5 blue-border and blue-filled circles of different size on a new plot
circles(
 x = sample(1:100, 10),
 y = sample(1:100, 10),
 radius = sample(1:10, 10),
 center = TRUE,
 border = c(rep("red", 5), rep("blue", 5)),
 col = c(rep(adjustcolor("firebrick", alpha = 0.2), 5), rep(adjustcolor("lightblue", alpha = 0.2), 5)),
 Res = 500,
 lwd = 1.5,
 lty = c(rep(1,5), rep(2,5))
)

}
\seealso{
\code{\link{plot}}, \code{\link{points}}, \code{\link{points}}
}
\author{
Quentin PETITJEAN
}
