% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2list.R
\name{convert2List}
\alias{convert2List}
\title{Convert a tracklets object to list of variables.}
\usage{
convert2List(trackDat)
}
\arguments{
\item{trackDat}{An object of class "tracklets" containing a list of tracklets and their characteristics classically used for further computations (at least x.pos, y.pos, frame).}
}
\value{
An object of class varList corresponding to a list of variables (vectors) retrieved from the tracklets.
}
\description{
Given an object of class "tracklets" containing a list of tracklets,
this function concatenate the tracklets data frame into a list based on the variables (varList object)
included within each tracklet. It also add the identity of the tracklets as a new variable named "trackletId".
}
\examples{

set.seed(2023)
# generate some dummy tracklets
## start to specify some parameters to generate tracklets
TrackN <- 3 # the number of tracklet to simulate
TrackL <-
  1:1000 # the length of the tracklets or a sequence to randomly sample tracklet length
id <- 0
TrackList <- MoveR::trackletsClass(stats::setNames(lapply(lapply(seq(TrackN), function(i)
  trajr::TrajGenerate(sample(TrackL, 1), random = TRUE, fps = 1)), function(j) {
    id <<- id + 1
    data.frame(
      x.pos = j$x - min(j$x),
      y.pos = j$y - min(j$y),
      frame = j$time,
     identity = paste("Tracklet", id, sep = "_")
    )
  }), seq(TrackN)))

# convert the list of tracklets to a simple list of variables
 trackDatList <- MoveR::convert2List(TrackList)
 str(trackDatList)

}
\seealso{
\code{\link{convert2Tracklets}}
}
\author{
Quentin PETITJEAN
}
