% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2Tracklets.R
\name{convert2Tracklets}
\alias{convert2Tracklets}
\title{Convert a list of variables to list of tracklets.}
\usage{
convert2Tracklets(trackDatList, by = NULL)
}
\arguments{
\item{trackDatList}{A list of vector corresponding to the variable of tracking data.}

\item{by}{A character vector identifying tracklets to join by.}
}
\value{
A list of data frames corresponding the the tracking data for each tracklets.
}
\description{
Given a list containing vectors corresponding to the various variable of tracking data
the function returns a list of data frames corresponding to the data for each tracklet based on tracklets identity.
}
\examples{

# generate some dummy tracklets
## start to specify some parameters to generate tracklets
Fragn <- 10 # the number of tracklet to simulate
FragL <- 1:1000 # the length of the tracklets or a sequence to randomly sample tracklet length

fragsList <- stats::setNames(lapply(lapply(seq(Fragn), function(i)
 trajr::TrajGenerate(sample(FragL, 1), random = TRUE, fps = 1)), function(j)
   data.frame(
     x.pos = j$x - min(j$x),
     y.pos = j$y - min(j$y),
     frame = j$time
   )), seq(Fragn))

# convert fragList to a simple list based on the variables present within each tracklets' data frame
trackDatList <- convert2List(fragsList)

# convert the list of variable to a liost of tracklet based on tracklets identity
trackDat <- convert2Tracklets(trackDatList, by = "fragsId")

}
\seealso{
\code{\link{convert2List}}
}
\author{
Quentin PETITJEAN
}
