% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutTracklets.R
\name{cutTracklets}
\alias{cutTracklets}
\title{Cut tracklets.}
\usage{
cutTracklets(trackDat, customFunc)
}
\arguments{
\item{trackDat}{A list of data frames containing tracking information for each tracklet or a data frame containing tracking information for one given tracklet.}

\item{customFunc}{A function used to cut/subset the tracklets.}
}
\value{
A list of data frames containing subsetted tracking information for each tracklet according to
the conditions specified by the customFunc argument.
}
\description{
Given a list of data frames containing tracking information for each tracklet the function
returns the tracklet part(s) that meet the condition specified in the customFunc.
}
\examples{

set.seed(2023)
# generate some dummy tracklets
## start to specify some parameters to generate tracklets
TrackN <- 20 # the number of tracklet to simulate
TrackL <- 1:1000 # the length of the tracklets or a sequence to randomly sample tracklet length

TrackList <- stats::setNames(lapply(lapply(seq(TrackN), function(i)
  trajr::TrajGenerate(sample(TrackL, 1), random = TRUE, fps = 1)), function(j)
    data.frame(
      x.pos = j$x - min(j$x),
      y.pos = j$y - min(j$y),
      frame = j$time
    )), seq(TrackN))

# example 1 :
## cut the tracklets by keeping only the part drawn during the first 25 seconds of the video
### create a time sequence from 0 to 25 seconds (25 seconds * 25 frames)
TimeSeq <- seq(0, 25 * 1, by = 1) 
### cut the tracklets
trackDatSub <- MoveR::cutTracklets(TrackList,
                      customFunc = function(x)
                      x[["frame"]] \%in\% TimeSeq)

par(mfrow = c(1,2))
### draw the tracklets before the cut
MoveR::drawTracklets(TrackList,
                 cex.leg = 0.8,
                 main = "")

### draw the tracklets after the cut (keep only the first 25 seconds of the video)
MoveR::drawTracklets(trackDatSub,
                 cex.leg = 0.8,
                 main = "")

}
\author{
Quentin PETITJEAN
}
