% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2Edge.R
\name{dist2Edge}
\alias{dist2Edge}
\title{Compute distance to the edge of an object (e.g., the arena) along a trajectory}
\usage{
dist2Edge(df, edge, customFunc)
}
\arguments{
\item{df}{A data frame containing at x, y coordinates named "x.pos", "y.pos", for a fragment}

\item{edge}{A data frame containing at x, y coordinates named "x.pos", "y.pos" and specifiyng the location of the
arena or any object edge}

\item{customFunc}{A function used to specify the formula used to compute the
distance between a given object or arena border and individuals along a trajectory
It is possible to call already implemented methods for Circular arena by calling customFunc = "CircularArena"}
}
\value{
this function returns a vector containing the distance between each points of the
trajectory and the closest point of the object edge
}
\description{
Given a data frames containing tracking information for a given fragment,
this function compute the euclidean distance between points defining the edge of an object (e.g., the arena)
and coordinates of the trajectory, the function returns the distance between each points of the
trajectory and the closest point of the object edge
}
\examples{

#TODO

}
