% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlSampleDat.R
\name{dlSampleDat}
\alias{dlSampleDat}
\title{Download sample data for MoveR package}
\usage{
dlSampleDat(dataSet = c(1, 2), dir = NA)
}
\arguments{
\item{dataSet}{A numeric value, either 1 or 2 corresponding to the desired sampled dataset.}

\item{dir}{The path to the directory where the sample dataset is saved, by default, the function will save it in a temporary directory (optional).}
}
\value{
This function returns a vector containing the path to the sample dataset (raw tracking data), the distance matrix from the border of the arena (performed using ImageJ) and the reference data, a .csv file containing particles location (manually detected) over several frames to perform sensitivity analyses.
}
\description{
Given the id of the dataset, this function download or retrieve the sample dataset used as example
to run the functions from the MoveR package.
}
\examples{

# Here we are selecting the first sample dataset.

dlSampleDat(dataSet = 1)

# If the "dir" argument is not specified, the data are saved in a temporary directory (deleted after the R session is properly terminated).
# In case the sample data has already been downloaded, the function will retrieve it instead of downlading it again.
# This behavior is similar if the "dir" argument is specified.

}
\author{
Quentin PETITJEAN
}
