% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLsampleData.R
\name{DLsampleData}
\alias{DLsampleData}
\title{Download sample data to depict the use of the MoveR package.}
\usage{
DLsampleData(
  dataSet = c(1, 2),
  tracker = c("Ctrax", "IdTracker", "TrackR", "TRex"),
  dir = NULL
)
}
\arguments{
\item{dataSet}{A numeric value, either 1 or 2 corresponding to the desired sampled dataset.}

\item{tracker}{A character string corresponding to the name of the tracking software from which data have to be downloaded: Ctrax, IdTracker, TrackR and/or TRex. By default the function returns the path
to sample data from each tracking software.}

\item{dir}{The path to the directory where the sample dataset is saved, by default, the function will save it in a temporary directory (optional).}
}
\value{
This function returns a vector containing the path to:
\itemize{
\item{the selected sample dataset (raw tracking data).}
\item{the distance matrix from the border of the arena (color tresholding - greyscale - performed using ImageJ).}
\item{the reference data (a .csv file containing particles location -manually detected- over several frames to perform sensitivity analyses).}
\item{the filtered/cleaned version of the tracking data (filtered using the approach described in the MoveR-Clean-FilterData vignette).}
}
Also, in case the sample dataset has already been downloaded in a temporary (default) or specified directory, the function retrieve the data instead of downloading it again.
}
\description{
Given the id of a sample dataset stored in \href{https://github.com/qpetitjean/MoveR_SampleData}{MoveR_SampleData Github repository},
this function download the selected sample dataset used as example to run the functions from the MoveR package and returns the local path to access the data.
}
\examples{
## Not run:

# Here we are selecting the first sample dataset, that have been tracked using \href{https://trex.run}{TRex}

DLsampleData(dataSet = 1, tracker = "TRex")

# If the "dir" argument is not specified, the data are saved in a temporary directory (deleted after the R session is properly terminated).
# In case the sample data has already been downloaded, the function will retrieve it instead of downloading it again.
# This behavior is similar if the "dir" argument is specified.

## End(Not run)
}
\author{
Quentin PETITJEAN
}
