% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorCheck.R
\name{.errorCheck}
\alias{.errorCheck}
\title{Check some conditions.}
\usage{
.errorCheck(
  trackDat = NA,
  df = NA,
  customFunc = NA,
  x.pos = NA,
  y.pos = NA,
  speedCol = NA,
  geometry = NA,
  minSpeed = NA,
  timeCol = NA,
  imgHeight = NA,
  Bstate = NA,
  pattern = NA,
  refDat = NA,
  turnAngle = NA,
  distTraveled = NA,
  Tstep = NA
)
}
\value{
this function returns an error message if some condition are not met; the function is not exported.
}
\description{
Given various arguments specific to the functions used in the MoveR package, the function test the condition and returns an error message accordingly.
}
\author{
Quentin PETITJEAN
}
\keyword{internal}
