% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBouts.R
\name{extactBouts}
\alias{extactBouts}
\title{extactBouts}
\usage{
extactBouts(trackDat, Bstate = NULL, pattern = NULL)
}
\arguments{
\item{trackDat}{A list of data frame containing tracking informations for each fragment}

\item{Bstate}{The name of a vector/df column present in each df of trackDat and indicating the state of the individual
along a fragment (i.e. numeric or characters)}

\item{pattern}{A character string containing a regular expression
to be matched in the Bstate vector. see gregexpr()}
}
\value{
this function returns a list containing several sublists:
1- a list containing the pattern specified by the user
2- a list containing the bouts the extracted according to the specified pattern, bouts are grouped and
named according to the id of the original fragment (e.g. the first bouts extracted from the first fragment will be located in the sublist named frags_1
and will be named frags_1.1)
}
\description{
Given a list of data frames containing tracking informations including a vector
containing behavioral patterns (e.g., behavioral states, location in areas) for each fragment,
the function extracts fragments part containing the specified pattern
}
\examples{

#TODO

}
