% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFunc.R
\name{filterFunc}
\alias{filterFunc}
\title{create and test custom filter for tracking data}
\usage{
filterFunc(trackDat, toFilter = NULL, customFunc = NULL)
}
\arguments{
\item{trackDat}{a list of data frame containing tracking infomations for each fragment
(e.g., maj.ax, angle, min.ax, x.pos, y.pos, ...)}

\item{toFilter}{an element of the fragment's data frame from which condition will be verified}

\item{customFunc}{a custom function containing condition(s) to be applied to toFilter argument}
}
\value{
a list of vector containing the result of a user defined condition test for each fragment
}
\description{
Given a list of data frame containing tracking informations for each fragment,
this function returns a list of vector containing the result of a user defined condition test for each fragment
}
\examples{

#' # Exemple 1: test for the presence of infinite value in x.pos, the result is TRUE 
  # if infinite values are detected
 
filterFunc(trackDat, toFilter = "x.pos", customFunc = function(x) is.infinite(x))


# Exemple 2: perform a condition test on the length of the individuals, the result is TRUE 
# if individual size is ranging between 1 and 20 pixels
 
filterFunc(trackDat, toFilter = "maj.ax", customFunc = function(x) x >= 1 & x <= 20)


}
