% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement_models.R
\name{flow_matrix}
\alias{flow_matrix}
\title{Population flow matrix}
\usage{
flow_matrix(populations, distances, params, model = "gravity",
  place_names = NULL, symmetric = TRUE)
}
\arguments{
\item{populations}{vector of populations.}

\item{distances}{distance matrix. This is assumed to be symmetric
so only lower triangle is of use.}

\item{params}{a named list of model parameters. For gravity model
the parameters are k, pow_from, pow_to, pow_dist.}

\item{model}{at the moment only gravity model is supported.}

\item{place_names}{optional. If not NULL, rownames of the returned
matrix will be set.}

\item{symmetric}{is the flow matrix symmetric. defaults to TRUE.}
}
\value{
N X N flow matrix where N is the length of the populations
vector. i,jth element in this matrix is the flow from location i
to location j. The diagonal is NA.
}
\description{
Population flow matrix
}
\details{
flow from a location to itself is undefined (NA) under
gravity model.
}
\author{
Sangeeta Bhatia
}
