% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement_models.R
\name{flow_vector}
\alias{flow_vector}
\title{Vector of population flow between locatuons}
\usage{
flow_vector(pop_from, pop_to, distances, params, model = "gravity")
}
\arguments{
\item{pop_from}{Numeric (optionally named) list of populations. If
the input is not named, integer names are assigned.}

\item{pop_to}{Numeric (optionally named) vector of populations}

\item{distances}{Numeric vector of distances. Element i is the
distance between location with pop_from[i] and pop_to[i].
Therefore the lengths of the two population vectors and the distance
vector should be the same}

\item{params}{parameters for the model. This should be a named
list with the names matching those required by the model. For
gravity model, these are: k, pow_from, pow_to, pow_dist.
\deqn{\phi(i, j) = k * n_i^pow_from * n_j^power_to / dist_{i, j}^pow_dist}}

\item{model}{Currently only gravity model is supported.}
}
\value{
a named list where the ith element is the flow from
(\code{pop_from})[i] to (\code{pop_to})[i].
is the flow between pop_from[i] and pop_to[j].
}
\description{
This function returns a vector of population flows under the
specified movement model with the given set of parameters.
}
\author{
Sangeeta Bhatia
}
