% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTime.R
\name{generateTime}
\alias{generateTime}
\title{Generate time}
\usage{
generateTime(trackingData)
}
\arguments{
\item{trackingData}{A list containing raw tracking datas}
}
\value{
A vector containing a sequence of frame number corresponding to the time serie
}
\description{
Given a vector containing the number of individuals detected in each frame (i.e., ntarget),
this function reconstruct times serie to recover actual time for replicated frame in the dataset
}
\examples{

# TODO

}
