% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement_models.R
\name{gravity_model}
\alias{gravity_model}
\title{Computes the flow from A to B under the gravity model}
\usage{
gravity_model(n_from, n_to, distance, params)
}
\arguments{
\item{n_from}{population of the source}

\item{n_to}{population of the destination}

\item{distance}{distance between the two places}

\item{params}{a named list of model parameters. For gravity model
the parameters are k, pow_from, pow_to, pow_dist.
destination}
}
\value{
estimated flow between source and destination
}
\description{
Given the populations of A and B and the distance between them,
return the estimated population flow between
them modeled as
\deqn{\phi(A,B) = K n_A^{\alpha}n_B^{\beta}/r_{AB}^{\gamma}}
}
\author{
Sangeeta Bhatia
}
