% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{import_IMDB}
\alias{import_IMDB}
\title{import_IMDB}
\usage{
import_IMDB(zip = 21210)
}
\arguments{
\item{zip}{zip code (5 numbers).}
}
\value{
a list of dataframes.
}
\description{
import_IMDB scrapes info of all the movies in theater that week.
}
