% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{import_movie}
\alias{import_movie}
\title{import_movie}
\usage{
import_movie(dat)
}
\arguments{
\item{dat}{webpage of movie on IMDB.}
}
\value{
a list of dataframes:
}
\description{
import_movie scrapes info of a movie on its IMDB page.
}
