% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_get.R
\name{list_get}
\alias{list_get}
\title{Get an object within a dataframe, a list or a nested list}
\usage{
list_get(my_list, to_find)
}
\arguments{
\item{my_list}{A dataframe, list or a nested list where to look for}

\item{to_find}{A character string corresponding to the name of the object to find}
}
\value{
The object to find or NULL whether the object have been not found
}
\description{
The function looks for an object name within all levels of a list and returns it when found
}
\examples{

# create some silly containers where to look for

df =  data.frame(values = c(1:4),
identity = LETTERS[1:4],
other = c("pif", "paf", "pouf","pof"))

slist = list(values = c(1:20),
identity = LETTERS[1:4],
other = c("pif", "paf", "pouf","pof"))

nlist = list(
values = c(1:20),
other = c("pif", "paf", "pouf","pof"),
nested = list(
 identity = LETTERS[1:4],
 values2 = c(1:20),
 other2 = c("pif", "paf", "pouf","pof")
))

nnlist = list(
values = c(1:20),
other = c("pif", "paf", "pouf","pof"),
nested = list(
 nested2 = list(identity = LETTERS[1:4]),
 values2 = c(1:20),
 other2 = c("pif", "paf", "pouf","pof")
))

# looking for "identity" containers within each element

list_get(df, "identity")
list_get(slist, "identity")
list_get(nlist, "identity")
list_get(nnlist, "identity")

}
