% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrorYFunc.R
\name{mirrorYFunc}
\alias{mirrorYFunc}
\title{Mirror Y coordinates}
\usage{
mirrorYFunc(YCoords, imgHeight = NA)
}
\arguments{
\item{YCoords}{A vector containing Y coordinates of detected individuals}

\item{imgHeight}{A numeric value expressed in pixels, the true length of Y axis
corresponding to the height of the image or video resolution (default = 1080)}
}
\value{
A vector containing new mirrored Y coordinates
}
\description{
Given a vector containing the Y coordinates of detected individuals and image height,
this function compute mirrored Y coordinates
}
\examples{

# TODO

}
