% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveR-package.R
\docType{package}
\name{MoveR-package}
\alias{MoveR}
\alias{MoveR-package}
\title{MoveR: R Package For Movement Analysis}
\description{
A set of functions helping users to process (i.e., filter, clean, draw and analyse) movement datas provided by 
    tracking software.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/qpetitjean/MoveR}
  \item Report bugs at \url{https://github.com/qpetitjean/MoveR/issues}
}

}
\author{
\strong{Maintainer}: Quentin PETITJEAN \email{q.petitjean1@gmail.com} (\href{https://orcid.org/0000-0003-2708-7831}{ORCID})

Authors:
\itemize{
  \item Vincent CALCAGNO \email{vincent.calcagno@inrae.fr} (\href{https://orcid.org/0000-0002-5781-967X}{ORCID})
  \item Silène LARTIGUE \email{Silene.Lartigue@inrae.fr}
  \item Mélina COINTE \email{melina.cointe@inrae.fr}
}

}
\keyword{internal}
