% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MovR-package.R
\docType{package}
\name{MovR-package}
\alias{MovR}
\alias{MovR-package}
\title{MovR: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several 
    lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/qpetitjean/MovR}
  \item Report bugs at \url{https://github.com/qpetitjean/MovR/issues}
}

}
\author{
\strong{Maintainer}: Quentin PETITJEAN \email{q.petitjean1@gmail.com} (\href{https://orcid.org/0000-0003-2708-7831}{ORCID})

}
\keyword{internal}
