% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternExtract.R
\name{patternExtract}
\alias{patternExtract}
\title{Extract specified patterns from fragments}
\usage{
patternExtract(trackDat, Bstate = NULL, pattern = NULL)
}
\arguments{
\item{trackDat}{A list of data frame containing tracking informations for each fragment, including a vector
containing behavioral patterns (e.g., behavioral states, location in areas).}

\item{Bstate}{The name of the column indicating the state of the individual along each fragment.}

\item{pattern}{A character string containing a regular expression to be matched in the Bstate vector. see gregexpr().}
}
\value{
This function returns a list containing the part of the fragments corresponding to the specified pattern.
The part of the fragments corresponding to the specified pattern are grouped into a list named according to the id of the original fragment
(e.g. the first detected pattern extracted from the first fragment is located in the sublist named frags_1 and is named frags_1.1).
}
\description{
Given a list of data frames containing tracking informations including a vector
containing behavioral patterns (e.g., behavioral states, location in areas) for each fragment,
the function extracts fragments part that contains the specified pattern.
}
\examples{

# TODO

}
