% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{polygons}
\alias{polygons}
\title{Generate points located on polygon vertices and draw polygon(s).}
\usage{
polygons(
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  sides = NULL,
  rotation = 0,
  center = TRUE,
  col = NULL,
  border = "black",
  lwd = 1,
  lty = 1,
  draw = TRUE
)
}
\arguments{
\item{x}{A single numeric value or a vector corresponding to the x value of the center of polygons}

\item{y}{A single numeric value or a vector corresponding to the y value of the center of polygons}

\item{width}{A single numeric value or a vector specifying the width of the polygons to generate.}

\item{height}{A single numeric value or a vector specifying the height of the polygons to generate.}

\item{sides}{A single numeric value or a vector specifying the number of sides belonging to the polygons to generate.}

\item{rotation}{A single numeric value or a vector specifying the angle of rotation of the polygons in radians.}

\item{center}{Either TRUE or a single integers or a vector of integers specifying the symbol(s) or a single character to be used to represent the polygons center
(see \code{\link[graphics]{points}} for possible values and their interpretation).}

\item{col}{The color or a vector of colors for filling the polygons, the default leaves polygons unfilled.}

\item{border}{The color or a vector of colors to draw the border of the polygons (default = "black").}

\item{lwd}{The value of line width or a vector of line width of the polygons border (default = 1).}

\item{lty}{An integer or a vector of integer corresponding to line type to be used for drawing polygons border, as in par (default = 1).}

\item{draw}{A logical value (i.e., TRUE or FALSE) indicating whether the polygons should be drawn or not (default = TRUE).}
}
\value{
This function returns a list containing the coordinates of the points generated on contour of the polygon(s) (i.e., vertices). Each element of the list
correspond to a polygon (an ROI). In case only one polygon is generated the function returns a dataframe. The function can also draw the circles on an existing plot window or, if there is no active plot window, on a new plot window.
}
\description{
Given x and y coordinates of the polygon(s)' center as well as width, height, number of sides, and rotation angle, this function returns a list containing the coordinates of the polygon(s) vertices and draw it.
}
\examples{

 set.seed(2023)
 plot(NULL, xlim = c(1,120), ylim = c(1,120), xlab = "x", ylab = "y")
 # draw 4 red-border and red-filled polygons of different shape, size and angle on a new plot
 polygonsCoords1 <- MoveR::polygons(
   x = sample(1:100, 4),
   y = sample(1:100, 4),
   width  = sample(5:40, 4),
   height = sample(5:40, 4),
   sides = sample(3:6, 4),
   rotation = sample(seq(0, 2*pi, 0.5), 4),
   center = TRUE,
   border = "red",
   col = adjustcolor("firebrick", alpha = 0.2),
   lwd = 1.5,
   lty = 1,
   draw = TRUE
 )
 str(polygonsCoords1)
 
 # draw 2 red-border and red-filled circles and 2 blue-border and blue-filled circles of different size on a new plot
 plot(NULL, xlim = c(-10,120), ylim = c(-10,120), xlab = "x", ylab = "y")
 polygonsCoords2 <- MoveR::polygons(
   x = sample(1:100, 4),
   y = sample(1:100, 4),
   width  = sample(5:40, 4),
   height = sample(5:40, 4),
   sides = sample(3:6, 4),
   rotation = sample(seq(0, 2*pi, 0.5), 4),
   center = TRUE,
   border = c(rep("red", 2), rep("blue", 2)),
   col = c(rep(adjustcolor("firebrick", alpha = 0.2), 2), rep(adjustcolor("lightblue", alpha = 0.2), 2)),
   lwd = 1.5,
   lty = c(rep(1, 2), rep(2, 2)),
   draw = TRUE
 )
str(polygonsCoords2)

}
\seealso{
\code{\link[graphics]{circles}}, \code{\link[graphics]{locPos}}, \code{\link[graphics]{locROI}}, \code{\link[graphics]{points}}
}
\author{
Quentin PETITJEAN
}
