% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCtrax.R
\name{readCtrax}
\alias{readCtrax}
\title{Import Ctrax tracking output .mat file.}
\usage{
readCtrax(ctraxPath, flipY = FALSE, imgHeight = NULL)
}
\arguments{
\item{ctraxPath}{The full path of the Ctrax output file (.mat).}

\item{flipY}{A logical value (i.e., TRUE or FALSE) indicating whether the origin of y coordinates should be mirrored. If TRUE, y coordinates are mirrored to start on the top-left (default = FALSE).}

\item{imgHeight}{A numeric value expressed in pixels, the length of Y axis
corresponding to the height of the image or video resolution (optional, only used when flipY = TRUE).}
}
\value{
A list containing 9 elements classically used for further computations.
Also, by default the function returns y coordinates starting on the bottom-left.
}
\description{
Given the path of the .mat file corresponding to Ctrax output,
this function returns a list of 9 vectors classically used for further computations using MoveR package:
\itemize{
\item{'maj.ax': }{the length of the major axis (i.e., the midline) for a particle over frame (i.e., length of the ellipse).}
\item{'angle': }{the particle's absolute angle in radians, orientation of the particle according to y-axis.}
\item{'min.ax': }{the length of the minor axis for a particle over frame (i.e., width of the ellipse).}
\item{'x.pos': }{x position of the particle's centroid.}
\item{'y.pos': }{y position of the particle's centroid.}
\item{'identity': }{the particle's identity given by the tracking software.}
\item{'frame': }{the video frame number at which the measurements has been made.}
\item{'ntargets': }{the number of particle tracked over each frame.}
\item{'timestamps': }{the elapsed time over each frame, in seconds.}
}

Also, the function can mirror y coordinates (see flipY argument).
}
\examples{

# Download the first dataset from the sample data repository
Path2Data <- MoveR::DLsampleData(dataSet = 1, tracker = "Ctrax")
Path2Data

# Import the list containing the 9 vectors classically used for further computation
# and do not mirror Y coordinates to start on the bottom-left
Data <- MoveR::readCtrax(Path2Data[[1]])
str(Data)
                         
}
\references{
Branson, K., Robie, A., Bender, J. et al. High-throughput ethomics in large groups of Drosophila. Nat Methods 6, 451–457 (2009). https://doi.org/10.1038/nmeth.1328.
\href{https://ctrax.sourceforge.net/}{Ctrax: The Caltech Multiple Walking Fly Tracker}
}
\seealso{
\code{\link{readTrackR}}, \code{\link{readTrex}}, \code{\link{readIdtracker}}, \code{\link{flipYCoords}}
}
\author{
Quentin PETITJEAN, Vincent CALCAGNO
}
