% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIdtracker.R
\name{readIdtracker}
\alias{readIdtracker}
\title{Load idtracker tracking output .npy file}
\usage{
readIdtracker(IdtrackerPath, mirrorY = FALSE, imgHeight = NA)
}
\arguments{
\item{IdtrackerPath}{The path of the idtrackerai output file (.npz) to load within R environment
(e.g. "C:/Users/\link{username}/Desktop/video_folder/trajectories.npz")}

\item{mirrorY}{TRUE or FALSE, set the origin of Y coords, if TRUE Y coords are mirrored}

\item{imgHeight}{A numeric value expressed in pixels, the true length of Y axis
corresponding to the height of the image or video resolution (default = 1080)}
}
\value{
A list containing tracking data
}
\description{

}
\examples{

# TODO

}
\seealso{
\code{\link{readCtrax}}, \code{\link{readTrex}}, \code{\link{readTrackR}}, \code{\link{mirrorYFunc}}
}
