% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrackR.R
\name{readTrackR}
\alias{readTrackR}
\title{Load TrackR tracking output .csv file}
\usage{
readTrackR(trackRPath, mirrorY = FALSE, imgHeight = NA)
}
\arguments{
\item{trackRPath}{The path of the TrackR output file (.csv) to load within R environment
(e.g. "C:/Users/\link{username}/Desktop/video_folder/Results.csv")}

\item{mirrorY}{TRUE or FALSE, set the origin of Y coords, if TRUE Y coords are mirrored}

\item{imgHeight}{A numeric value expressed in pixels, the true length of Y axis
corresponding to the height of the image or video resolution (optional, only used when mirrorY = TRUE)}
}
\value{
A list containing tracking data
}
\description{
Given a .csv file corresponding to TrackR tracking output,
this function process the .csv file to return a list containing tracking data
If needed this function can also mirror Y coordinates
}
\examples{

# TODO

}
\seealso{
\code{\link{readCtrax}}, \code{\link{readTrex}}, \code{\link{readIdtracker}}, \code{\link{mirrorYFunc}}
}
