% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamplTracklets.R
\name{resamplTracklets}
\alias{resamplTracklets}
\title{Resampling tracklets.}
\usage{
resamplTracklets(trackDat, TimeCol = NULL, Tstep = NULL, progress = TRUE)
}
\arguments{
\item{TimeCol}{A character string corresponding to the name of the column containing time information (e.g., "frame").}

\item{Tstep}{A numeric value corresponding to the length of the resampling step according to the time unit used in TimeCol (e.g., frame).}

\item{progress}{A logical value (i.e., TRUE or FALSE) indicating whether a progress bar should be displayed to inform process progression.}

\item{trackDatA}{A list of data frame containing tracking informations for each tracklet.}
}
\value{
A list of data frame containing the resampled tracklets.
}
\description{
Given a list of data frame containing tracking informations for each tracklet,
this function returns a list of resampled tracklets according to the length of the resampling
step specified by the Tstep argument.
}
\examples{

set.seed(2023)
# generate some dummy tracklets
## start to specify some parameters to generate tracklets
trackN <- 50 # the number of tracklets to simulate
trackL <-
  100:1000 # the length of the tracklets or a sequence to randomly sample tracklets length

trackList <- stats::setNames(lapply(lapply(seq(trackN), function(i)
  trajr::TrajGenerate(sample(trackL, 1), random = TRUE, fps = 1)), function(j)
    data.frame(
      x.pos = j$x - min(j$x),
      y.pos = j$y - min(j$y),
      frame = j$time
    )), seq(trackN))

# check the tracklets
MoveR::drawTracklets(trackList,
                 timeCol = "frame")

# convert the time unit (frame) to seconds using analyseTracklets (according to a frame rate of 25 fps)
trackListV1 <-
  MoveR::analyseTracklets(trackList,
                      customFunc = list(
                        # convert the time expressed in frame in second using a conversion factor of 25 frame per second
                        TimeSec = function(x)
                          x[["frame"]] / 25
                      ))
# example 1: resample the tracklets every 0.2 seconds
## check the size of the tracklets
trackSize <- unlist(lapply(trackListV1, function(x)
  nrow(x)))

## resample the tracklets every 1 seconds
trackListSampled1S <- MoveR::resamplTracklets(trackListV1,
                                           TimeCol = "TimeSec",
                                           Tstep = 1)

## check the size of the tracklets after resampling
trackSize1s <- unlist(lapply(trackListSampled1S, function(x)
  nrow(x)))

## Compare the tracklets size
cbind(trackSize, trackSize1s)

# example 2: resample the tracklets every 1 minutes

## resample the tracklets every 0.10 seconds
trackListSampled0.10S <- MoveR::resamplTracklets(trackListV1,
                                              TimeCol = "TimeSec",
                                              Tstep = 0.10)
## here some time step are not found in the tracklets' list (e.g., 0.10, 0.30),
## they are hence replaced by NA in the output list

## check the size of the tracklets after resampling
trackSize0.10S <-
  unlist(lapply(trackListSampled0.10S, function(x)
    nrow(x)))

## Compare the tracklets size
cbind(trackSize, trackSize1s, trackSize0.10S)

}
\author{
Quentin PETITJEAN
}
