% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim1_univariate.R
\name{run_sim1_univariate}
\alias{run_sim1_univariate}
\title{A Function to fit univariate ABM for simulation 1}
\usage{
run_sim1_univariate(num_mcmc, num_particles, data, m_mu, sigmasq_m, nu,
  sigmasq0, nu_eps, sigmasq0_eps, mu_theta_mean, var_theta)
}
\arguments{
\item{num_mcmc}{The number of MCMC iterations}

\item{num_particles}{The number of particles to use in particle filter}

\item{data}{Time_points x 2 matrix of observed locations}

\item{m_mu}{Prior values for step size mean: # mu ~ N(m_mu, sigmasq_m)}

\item{sigmasq_m}{Prior values for step size mean: # mu ~ N(m_mu, sigmasq_m)}

\item{nu}{Prior values for step size standard deviation: sigmasq_u ~ IG(nu/2,sigmasq0 * nu / 2)}

\item{sigmasq0}{Prior values for step size standard deviation: sigmasq_u ~ IG(nu/2,sigmasq0 * nu / 2)}

\item{nu_eps}{Prior values for observation error: # sigmasq_eps ~ IG(nu_eps/2,nu_eps * sigmasq0_eps / 2)}

\item{sigmasq0_eps}{Prior values for observation error: # sigmasq_eps ~ IG(nu_eps/2,nu_eps * sigmasq0_eps / 2)}

\item{mu_theta_mean}{Prior values for projected normal: mu_theta ~ N(mu_theta_mean, var_theta)}

\item{var_theta}{Prior values for projected normal: mu_theta ~ N(mu_theta_mean, var_theta)}
}
\value{
A list containing the path and the log probability of the path
}
\description{
A Function to fit univariate ABM for simulation 1
}
