% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim2.R
\name{run_sim2}
\alias{run_sim2}
\title{A Function to fit multistate ABM for simulation 2}
\usage{
run_sim2(num_mcmc, num_particles, data, m_mu1, m_mu2, sigmasq_m1,
  sigmasq_m2, nu, sigmasq01, sigmasq02, nu_eps, sigmasq0_eps,
  mu_theta_mean, var_theta)
}
\arguments{
\item{num_mcmc}{The number of MCMC iterations}

\item{num_particles}{The number of particles to use in particle filter}

\item{data}{num_agents X time_points x 2 array of observed locations}

\item{m_mu1}{Prior values for step size mean for state 1: # mu ~ N(m_mu, sigmasq_m)}

\item{m_mu2}{Prior values for step size mean for state 2: # mu ~ N(m_mu, sigmasq_m)}

\item{sigmasq_m1}{Prior values for step size variance for state 1: # mu ~ N(m_mu, sigmasq_m)}

\item{sigmasq_m2}{Prior values for step size variance for state 2: # mu ~ N(m_mu, sigmasq_m)}

\item{nu}{Prior values for step size standard deviation: sigmasq_u ~ IG(nu/2,sigmasq0 * nu / 2)}

\item{sigmasq01}{Prior values for step size standard deviation for state 1: sigmasq_u ~ IG(nu/2,sigmasq0 * nu / 2)}

\item{sigmasq02}{Prior values for step size standard deviation for state 2: sigmasq_u ~ IG(nu/2,sigmasq0 * nu / 2)}

\item{nu_eps}{Prior values for observation error: # sigmasq_eps ~ IG(nu_eps/2,nu_eps * sigmasq0_eps / 2)}

\item{sigmasq0_eps}{Prior values for observation error: # sigmasq_eps ~ IG(nu_eps/2,nu_eps * sigmasq0_eps / 2)}

\item{mu_theta_mean}{Prior values for projected normal: mu_theta ~ N(mu_theta_mean, var_theta)}

\item{var_theta}{Prior values for projected normal: mu_theta ~ N(mu_theta_mean, var_theta)}
}
\value{
A list containing the path and the log probability of the path
}
\description{
A Function to fit multistate ABM for simulation 2
}
