% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidWindow.R
\name{slidWindow}
\alias{slidWindow}
\title{Sliding window.}
\usage{
slidWindow(x, Tstep, customFunc)
}
\arguments{
\item{x}{A vector containing data to smooth.}

\item{Tstep}{A numeric value corresponding to the length of the sliding window (i.e., the number of values).}

\item{customFunc}{A function used to perform the computation along the sliding window.}
}
\value{
A vector containing smoothed data.
}
\description{
Given a vector containing numeric data, this function returns a vector of smoothed data
by performing the specified computation along a centered sliding window.
}
\examples{

set.seed(123) 
# compute smoothed mean 
smoothMean <- slidWindow( x = rnorm(150), Tstep = 15, customFunc = function (x) mean(x, na.rm = T))
# compute smoothed variance 
smoothVar <- slidWindow( x = rnorm(150), Tstep = 15, customFunc = function (x) var(x, na.rm = T))
# compute smoothed sd
smoothSd <- slidWindow( x = rnorm(150), Tstep = 15, customFunc = function (x) sd(x, na.rm = T))

}
\author{
Quentin PETITJEAN
}
