% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortVertices.R
\name{sortVertices}
\alias{sortVertices}
\title{Sort the vertices of an ROI in a clockwise order}
\usage{
sortVertices(ROI)
}
\arguments{
\item{ROI}{A dataframe or a matrix containing x and y coordinates corresponding to the vertices (or any points located on the edge) of an ROI (i.e., polygon).}
}
\value{
This function returns a dataframe containing the sorted coordinates of the ROI's vertices in a clockwise order.
}
\description{
Given a dataframe or a matrix containing the x and y coordinates corresponding to the vertices or any points located on the edge of an ROI (i.e., polygon),
this function returns the vertices sorted in a clockwise order according to their angles.
}
\examples{

set.seed(2023)
# generate dummy vertices for polygon
ROI <- data.frame(x = sample(1:100, 6), y = sample(1:100, 6))
par(mfrow=c(1,2))
# plot the polygon without ordering the vertices
plot(NULL, xlim = c(1,100), ylim = c(1,100))
polygon(ROI)
# plot the polygon with ordered vertices
plot(NULL, xlim = c(1,100), ylim = c(1,100))
polygon(sortVertices(ROI))

}
\author{
Quentin PETITJEAN
}
