% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateSeq.R
\name{stateSeq}
\alias{stateSeq}
\title{Extract specified patterns/sequences from fragments.}
\usage{
stateSeq(trackDat, Bstate = NULL, pattern = NULL, perl = FALSE, fixed = FALSE)
}
\arguments{
\item{trackDat}{A list of data frame containing tracking informations for each fragment, including a vector
containing behavioral patterns (e.g., behavioral states, location in areas).}

\item{Bstate}{The name of the column indicating the state of the particle along each fragment.}

\item{pattern}{A character string containing a regular expression to look for in the Bstate vector (see \code{\link[base]{grep}}).}

\item{perl}{A logical value indicating whether Perl-like regular expression should be used (default = FALSE, see \code{\link[base]{regex}}).}

\item{fixed}{A logical value indicating whether a literal regular expression (exact match) should be used (default = FALSE, see \code{\link[base]{regex}}).}
}
\value{
This function returns a list containing the part of the fragments corresponding to the specified pattern.
The part of the fragments corresponding to the specified pattern are grouped into a list named according to the id of the original fragment
(e.g. the first detected pattern extracted from the first fragment is located in the sublist named frags_1 and is named frags_1.1).
}
\description{
Given a list of data frames containing tracking informations including a vector
containing behavioral patterns (e.g., behavioral states, location in areas) for each fragment,
the function detect and extracts the fragment parts that contains the specified pattern.
}
\examples{

# TODO

}
\author{
Quentin PETITJEAN
}
