% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3SubsetFunc.R
\name{[.varList}
\alias{[.varList}
\title{Subsetting varList or tracklets objects.}
\usage{
\method{[}{varList}(x, i, ..., drop = FALSE)
}
\value{
these functions return the subsetted varList or tracklets objects
}
\description{
use of \code{[} to subset varList or tracklets objects
}
\author{
Quentin PETITJEAN
}
\keyword{internal}
