% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackletsClass.R
\name{trackletsClass}
\alias{trackletsClass}
\title{Create an object of class \code{tracklets}.}
\usage{
trackletsClass(tracklets)
}
\arguments{
\item{tracklets}{A list of data frame containing tracklets information. The presence of x, y coordinates and time column named as x.pos, y.pos and frame is necessary for each tracklets.}
}
\value{
An object of class \code{tracklets}.
}
\description{
create an object of class \code{tracklets} from a list of tracklets.
}
\examples{

set.seed(2023)
# generate some dummy tracklets
## start to specify some parameters to generate tracklets
TrackN <- 500 # the number of tracklet to simulate
TrackL <- 1:1000 # the length of the tracklets or a sequence to randomly sample tracklet length

TrackList <- stats::setNames(lapply(lapply(seq(TrackN), function(i)
  trajr::TrajGenerate(sample(TrackL, 1), random = TRUE, fps = 1)), function(j)
    data.frame(
      x.pos = j$x - min(j$x),
      y.pos = j$y - min(j$y),
      frame = j$time
    )), seq(TrackN))
    
 test <- trackletsClass(TrackList)
 
}
\author{
Quentin PETITJEAN
}
