% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackSensitiv.R
\name{trackSensitiv}
\alias{trackSensitiv}
\title{trackSensitiv}
\usage{
trackSensitiv(referenceDf, trackingData, radius = 20, imgRes = c(1920, 1080))
}
\arguments{
\item{referenceDf}{a list of data frame containing "true" x and y coordinates of individuals (manually detected using imageJ for instance)
as well as the id of the frame. In case the list contains several data frame (i.e., location of individuals for several frames),
sensitivity index is averged and Sd, Se and n are returned}

\item{trackingData}{A list of tracking fragments}

\item{radius}{a numeric value expressed in pixels and corresponding to the radius of the circle used to determine whether values
are considered similar to those within the referenceDF or not (default = 20)}

\item{imgRes}{A vector of 2 numeric values, resolution of the video used as x and y limit to create a caneva
(i.e., the number of pixels in image width and height, e.g., 1920 x 1080). (default = c(1920, 1080))}
}
\value{
a set of nested lists containing the summary of the analysis:
list 1: sensitivity index, Sd, Se, n (Sd, Se are only computed if the list of referenceDf is > 1)
list 2: a data frame containing detailed sensitivity index and frame on which test have been performed
list 3: the list of the False negative
list 4: the list of the false positive
}
\description{
compare the location of individual performed manually and by tracking software to return false positive, false negative and
sensitivity index
}
\examples{

#TODO

}
\seealso{

}
