% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackStats.R
\name{trackStats}
\alias{trackStats}
\title{tracking statistics}
\usage{
trackStats(trackDat, frameR = NA, scale = NA, unit = "unit")
}
\arguments{
\item{trackDat}{A list of data frame containing tracking informations for each fragment
(e.g., maj.ax, angle, min.ax, x.pos, y.pos, ...)}

\item{frameR}{A numeric value expressed in frames per second - the frame rate corresponding to
the frequency at which frames are recorded/displayed in the video}

\item{scale}{A ratio corresponding to the scaling factor to be applied to the trajectory coordinates
(e.g., size in cm / size in pixels; see TrajScale())}

\item{unit}{The unit expected after scaling (e.g., "cm", "m", ...)}
}
\value{
A list of basic metrics about video and fragments
}
\description{
Given a list of data frame containing tracking informations for each fragment,
this function compute and return various basic metrics about video and fragments
}
\examples{

# TODO

}
