% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnAngle.R
\name{turnAngle}
\alias{turnAngle}
\title{Compute turning angle}
\usage{
turnAngle(df, unit = c("radians", "degrees"))
}
\arguments{
\item{df}{A data frame containing x, y coordinates columns named "x.pos", "y.pos" for a given fragment}

\item{unit}{A character string indicating whether the function should returns turning angle in radians or degrees}
}
\value{
this function returns a vector containing the value of turning angle computed along a given fragment
}
\description{
Given a data frames containing tracking informations for a given fragment,
this function compute turning angle of individual and returns a vector containing the value of
turning angle along this fragment
}
\examples{

#TODO

}
