% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_valueBoxRow.R
\name{valueBoxRow}
\alias{valueBoxRow}
\title{valueBoxRow}
\usage{
valueBoxRow(outputId)
}
\arguments{
\item{outputId}{list of outputId}
}
\description{
creates a row of valueBox with a single value
}
\keyword{internal}
