% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.R
\name{circles}
\alias{circles}
\title{Generate points located on circles contour and draw circle(s).}
\usage{
circles(
  x = NULL,
  y = NULL,
  radius = NULL,
  Res = 500,
  center = NULL,
  col = NULL,
  border = "black",
  lwd = 1,
  lty = 1,
  draw = TRUE
)
}
\arguments{
\item{x}{A coordinate vector of circles to plot.}

\item{y}{A coordinate vector of circles to plot.}

\item{radius}{A single numeric value or a vector specifying the radius of the circles to plot.}

\item{Res}{A numeric value corresponding to the resolution of the circles (i.e., the number of points used to draw the circles contour, default = 500).}

\item{center}{Either TRUE or a single integers or a vector of integers specifying the symbol(s) or a single character to be used to represent the circle center
(see \code{\link[graphics]{points}} for possible values and their interpretation).}

\item{col}{The color or a vector of colors for filling the circles, the default leaves polygons unfilled.}

\item{border}{The color or a vector of colors to draw the border of the circles (default = "black").}

\item{lwd}{The value of line width or a vector of line width of the circles border (default = 1).}

\item{lty}{An integer or a vector of integer corresponding to line type to be used for drawing circles border, as in par (default = 1).}

\item{draw}{A logical value (i.e., TRUE or FALSE) indicating whether the circle(s) should be drawn or not (default = TRUE).}
}
\value{
Draw the circles from the coordinates of the center and a value of the circle radius on an existing plot window or, if there is no active plot window, on a new plot window.
It also returns a list containing the coordinates of the points simulated to draw the contour of the circle(s).
}
\description{
Given x and y coordinates of the circle center as well as the circle radius, this function draw circle(s) and returns a list containing the coordinates of the points simulated to draw the contour of the circle(s).
}
\examples{

set.seed(2023)
plot(NULL, xlim = c(1,120), ylim = c(1,120), xlab = "x", ylab = "y")
# draw 4 red-border and red-filled circles of different size on a new plot
circleCoords1 <- circles(
  x = sample(1:100, 4),
  y = sample(1:100, 4),
  radius = sample(5:15, 4),
  center = TRUE,
  border = "red",
  col = adjustcolor("firebrick", alpha = 0.2),
  Res = 500,
  lwd = 1.5,
  lty = 1,
  draw = TRUE
)
str(circleCoords1)

# draw 2 red-border and red-filled circles and 2 blue-border and blue-filled circles of different size on #' a new plot
plot(NULL, xlim = c(-10,120), ylim = c(-10,120), xlab = "x", ylab = "y")
circleCoords2 <- circles(
  x = sample(1:100, 4),
  y = sample(1:100, 4),
  radius = sample(5:20, 4),
  center = TRUE,
  border = c(rep("red", 2), rep("blue", 2)),
  col = c(rep(adjustcolor("firebrick", alpha = 0.2), 2), rep(adjustcolor("lightblue", alpha = 0.2), 2)),
  Res = 500,
  lwd = 1.5,
  lty = c(rep(1, 2), rep(2, 2)),
  draw = TRUE
)
str(circleCoords2)

}
\seealso{
\code{\link[graphics]{polygon}}, \code{\link[graphics]{points}}
}
\author{
Quentin PETITJEAN
}
