% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2frags.R
\name{convert2frags}
\alias{convert2frags}
\title{Convert raw tracking data to fragments list}
\usage{
convert2frags(tracking_data, by = NULL)
}
\arguments{
\item{tracking_data}{A list containing raw tracking datas}

\item{by}{A character vector identifying fragments to join by}
}
\value{
A list of fragments
}
\description{
Given a list containing raw tracking data,
this function convert it to a list of fragments
}
\examples{

# TODO

}
