% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2list.R
\name{convert2list}
\alias{convert2list}
\title{Convert fragments list to simple list}
\usage{
convert2list(tracking_data)
}
\arguments{
\item{tracking_data}{A list containing fragments tracking datas}
}
\value{
A list of tracking data as in raw file
}
\description{
Given a nested list containing fragments sublists
to a simple list as original raw data for use in track_stat
}
\examples{

# TODO

}
