% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2Tracklets.R
\name{convert2Tracklets}
\alias{convert2Tracklets}
\title{Convert a list of variables to list of tracklets.}
\usage{
convert2Tracklets(trackDatList, by = NULL)
}
\arguments{
\item{trackDatList}{A list of vector corresponding to the variable characterizing the tracking data.}

\item{by}{A character vector identifying the tracklets to join by.}
}
\value{
A list of data frames corresponding the tracking data for each tracklets.
}
\description{
Given a list containing vectors corresponding to the various variable of tracking data
the function returns a list of data frames corresponding to the data for each tracklet based on tracklets identity.
}
\examples{

set.seed(2023)
# generate some dummy tracklets
## start to specify some parameters to generate tracklets
TrackN <- 3 # the number of tracklet to simulate
TrackL <-
  1:1000 # the length of the tracklets or a sequence to randomly sample tracklet length
id <- 0
TrackList <- stats::setNames(lapply(lapply(seq(TrackN), function(i)
  trajr::TrajGenerate(sample(TrackL, 1), random = TRUE, fps = 1)), function(j) {
    id <<- id + 1
    data.frame(
      x.pos = j$x - min(j$x),
      y.pos = j$y - min(j$y),
      frame = j$time,
     identity = paste("Tracklet", id, sep = "_")
    )
  }), seq(TrackN))

 # convert the list of tracklets to a simple list of variables
 trackDatList <- MoveR::convert2List(TrackList)
 
 # convert back the list of variables to a list of tracklet based on tracklets identity
 trackDat <- MoveR::convert2Tracklets(trackDatList, by = "trackletId")
 str(trackDat) # display only the three first elements (dataframes) of the list

}
\seealso{
\code{\link{convert2List}}
}
\author{
Quentin PETITJEAN
}
