% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutFrags.R
\name{cutFrags}
\alias{cutFrags}
\title{CutFrags}
\usage{
cutFrags(trackDat, customFunc)
}
\arguments{
\item{trackDat}{A list of data frame containing tracking informations for each fragment (including a timeline)}

\item{customFunc}{A function used to cut/subset the fragments}
}
\value{
This function returns a list of data frames containing subsetted tracking informations for each fragments
that meet the conditions specified by customFunc
}
\description{
Given a list of data frames containing tracking informations for each fragments the function
returns the fragments part that meet the condition specified in the customFunc
}
\examples{

#TODO

}
