% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2Pt.R
\name{dist2Pt}
\alias{dist2Pt}
\title{Compute distance to the a given point}
\usage{
dist2Pt(df, obj)
}
\arguments{
\item{df}{A dataframe containing at least 2 values corresponding to x, y coordinates named "x.pos", "y.pos", for a fragment}

\item{obj}{A dataframe containing 2 values corresponding to x, y coordinates named "x.pos", "y.pos"
and specifying the location of an object or a point of interest}
}
\value{
this function returns a vector containing the distance between each points of the
trajectory and the specified point
}
\description{
Given a data frames containing tracking information for a given fragment,
this function compute the euclidean distance between a point and coordinates of the trajectory,
the function returns the distance between the specified point and the individual along the trajectory
}
\examples{

#TODO

}
