% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distTraveled.R
\name{distTraveled}
\alias{distTraveled}
\title{Compute distance to something along a trajectory}
\usage{
distTraveled(df, Step = 1)
}
\arguments{
\item{df}{A data frame containing at x, y coordinates named "x.pos", "y.pos", for a fragment}

\item{Step}{A numeric value specifying the number of unit between the two position for which distance will be computed}
}
\value{
this function returns a vector containing the distance between each points of the
trajectory according to the specified step
}
\description{
Given a data frames containing tracking information for a given fragment,
this function compute the euclidean distance along the trajectory according to a given step
}
\examples{

#TODO

}
