% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploredArea.R
\name{exploredArea}
\alias{exploredArea}
\title{Compute the total surface explored and display the corresponding heatmap.}
\usage{
exploredArea(
  trackDat,
  binRad = NULL,
  imgRes = c(NA, NA),
  scale = 1,
  timeCol = "frame",
  timeWin = list(c(0, Inf)),
  graph = TRUE,
  saveGraph = FALSE
)
}
\arguments{
\item{trackDat}{A list of data frame containing tracking information for each tracklet (including a timeline)
or a data frame containing tracking information for one tracklet.}

\item{binRad}{A numeric value, expressed in pixels, corresponding to the diameter of the typical surface a particle can "explore" around its position.}

\item{imgRes}{A vector of 2 numeric values, the resolution of the video used as x and y limit of the plot (i.e., the number of pixels in image width and height).
If imgRes is unspecified, the function retrieve it using x and y maximum values + 5\%.}

\item{scale}{A ratio corresponding to the scaling factor to be applied to the trajectory coordinates (optional).}

\item{timeCol}{A character string corresponding to the name of the column containing Time information (e.g., "frame").}

\item{timeWin}{A list of one or several vector containing 2 numeric values separated by a comma
corresponding to the time interval between which the tracklets have to be drawn according to timeCol (optional).}

\item{graph}{A logical value (i.e., TRUE or FALSE) indicating whether the various diagnostics plots should be displayed or not (default = TRUE).}

\item{saveGraph}{A logical value (i.e., TRUE or FALSE) indicating whether the heatmap should be saved as a .tiff file in the hardrive - within the working directory (default = FALSE).
Alternatively, the user can specify the saving function and hence the extension and the
directory where the heatmap should be saved (see for instance \code{\link[grDevices]{png}}, \code{\link[grDevices]{jpeg}} or \code{\link[grDevices]{tiff}}.}
}
\value{
This function returns the total surface explored in a geometrically consistent and scalable manner and if
graph argument is TRUE, it displays an heatmap (i.e., hexbin plot) of the explored areas.
}
\description{
Given a list of tracking tracklets containing cartesian coordinates, and a value of
reaction distance this function displays an heatmap (i.e., hexbin plot) of the explored areas and returns the
total surface explored in a geometrically consistent and scalable manner.
}
\examples{

set.seed(2023)
# generate some dummy tracklets
## start to specify some parameters to generate tracklets
TrackN <- 40 # the number of tracklet to simulate
TrackL <-
  1:1000 # the length of the tracklets or a sequence to randomly sample tracklet length
id <- 0
TrackList <- stats::setNames(lapply(lapply(seq(TrackN), function(i)
  trajr::TrajGenerate(sample(TrackL, 1), random = TRUE, fps = 1)), function(j) {
    id <<- id + 1
    data.frame(
      x.pos = j$x - min(j$x),
      y.pos = j$y - min(j$y),
      frame = j$time,
      identity = paste("Tracklet", id, sep = "_")
    )
  }), seq(TrackN))

# check the tracklets
MoveR::drawTracklets(TrackList,
                 timeCol = "frame")

# compute the total surface explored and displays the heatmap for all tracklets
# NB: Plot can be displayed and/or saved using graph and saveGraph arguments
MoveR::exploredArea(TrackList,
                    binRad = 20,
                    scale = 1,
                    timeCol = "frame")

# compute the surface explored for each tracklet and display (but do not save) the heatmap for the 2 first tracklets
# by combining exploredArea and analyseTracklets, the surface explored for each tracklet is appended to the data
# of the corresponding tracklets.
ExplorTest <- MoveR::analyseTracklets(TrackList[1:2],
                                  customFunc = list(
                                    exploredArea = function(x)
                                      MoveR::exploredArea(
                                        x,
                                        binRad = 20,
                                        scale = 1,
                                        timeCol = "frame"
                                      )
                                  ))
str(ExplorTest)

}
\author{
Quentin PETITJEAN
}
