% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploredArea.R
\name{exploredArea}
\alias{exploredArea}
\title{Explored Area}
\usage{
exploredArea(
  trackDat,
  binRad = NULL,
  imgRes = c(1920, 1080),
  scale = 1,
  timeCol = "frame",
  timeWin = list(c(0, Inf))
)
}
\arguments{
\item{trackDat}{A list of tracking fragments to plot along x and y coordinates}

\item{binRad}{A numeric value corresponding to the diameter of the typical surface a particle can "explore" around its position
For instance, the reaction distance of a Trichograms (i.e., a parasitoid micro-wasp) is about 4 mm, in this case, a reasonable value is of order 16 mm^2,
the diameter of such a cell is hence about 8 mm}

\item{imgRes}{A vector of 2 numeric values, resolution of the video used as x and y limit of the plot
(i.e., the number of pixels in image width and height, e.g., 1920 x 1080).}

\item{scale}{A ratio corresponding to the scaling factor to be applied to the trajectory coordinates (optional)}

\item{timeCol}{A character string corresponding to the name of the column containing Time information (e.g., "frame") (optional)}

\item{timeWin}{A list of one or several vector containing 2 numeric values separated by a comma
corresponding to the time interval between which the fragments have to be drawn according to timeCol (optional)}
}
\value{
displays an heatmap (i.e., hexbin plot) of the explored areas and returns the
total surface explored in a geometrically consistent and scalable manner
}
\description{
Given a list of tracking fragments containing cartesian coordinates, and a value of
reaction distance this function displays an heatmap (i.e., hexbin plot) of the explored areas and returns the
total surface explored in a geometrically consistent and scalable manner
}
\examples{



}
