% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFrags.R
\name{filterFrags}
\alias{filterFrags}
\title{apply custom filter on tracking data}
\usage{
filterFrags(trackDat, filter, splitCond = TRUE, minDur = 1)
}
\arguments{
\item{trackDat}{a list of data frame containing tracking informations for each fragment
(e.g., maj.ax, angle, min.ax, x.pos, y.pos, ...)}

\item{filter}{a list of vector as returned by filterFunc and containing the result of a condition test}

\item{splitCond}{the result of the condition test for which fragments have to be splitted (TRUE or FALSE)}

\item{minDur}{the minimum duration of the fragment to keep after splitting, expressed in frame}
}
\value{
this function returns a new list of dataframe corresponding to each new fragments created based on the result of filterFunc
}
\description{
Given a list of data frame containing tracking informations for each fragment and the result
of a specific condition test as returned by filterFunc this function returns a list of dataframes
corresponding to new fragments
}
\examples{

#TODO

}
\seealso{
\code{\link{filterFunc}}
}
