% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdStateSeq.R
\name{IdStateSeq}
\alias{IdStateSeq}
\title{Identify and extract specified patterns/sequences from tracklets trajectories.}
\usage{
IdStateSeq(
  trackDat,
  Bstate = NULL,
  pattern = NULL,
  perl = FALSE,
  fixed = FALSE
)
}
\arguments{
\item{trackDat}{A list of data frame containing tracking informations for each tracklet, including a vector
containing behavioral patterns (e.g., behavioral states, location in areas).}

\item{Bstate}{The name of the column indicating the state of the particle along each tracklet.}

\item{pattern}{A character string containing a regular expression to look for in the Bstate vector (see \code{\link[base]{grep}}).}

\item{perl}{A logical value (i.e., TRUE or FALSE) indicating whether Perl-like regular expression should be used (default = FALSE, see \code{\link[base]{regex}}).}

\item{fixed}{A logical value (i.e., TRUE or FALSE) indicating whether a literal regular expression (exact match) should be used (default = FALSE, see \code{\link[base]{regex}}).}
}
\value{
This function returns a list containing the part of the tracklets corresponding to the specified pattern.
The part of the tracklets corresponding to the specified pattern are grouped into a list named according to the id of the original tracklet
(e.g. the first detected pattern extracted from the first tracklet is located in the sublist named frags_1 and is named frags_1.1).
}
\description{
Given a list of data frames containing tracking informations including a vector
containing behavioral patterns (e.g., behavioral states, location in areas) for each tracklet,
the function identify and extract tracklets parts that contains the specified pattern.
}
\examples{

# TODO

}
\author{
Quentin PETITJEAN
}
