% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MovR-package.R
\docType{package}
\name{MovR-package}
\alias{MovR}
\alias{MovR-package}
\title{MovR: R Package For Movement Analysis}
\description{
A set of functions helping users to process (i.e., filter, clean, draw and analyse) movement datas provided by 
    tracking software.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/qpetitjean/MovR}
  \item Report bugs at \url{https://github.com/qpetitjean/MovR/issues}
}

}
\author{
\strong{Maintainer}: Quentin PETITJEAN \email{q.petitjean1@gmail.com} (\href{https://orcid.org/0000-0003-2708-7831}{ORCID})

Authors:
\itemize{
  \item Vincent CALCAGNO \email{vincent.calcagno@inrae.fr} (\href{https://orcid.org/0000-0002-5781-967X}{ORCID})
}

}
\keyword{internal}
