% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCtrax.R
\name{readCtrax}
\alias{readCtrax}
\title{Load Ctrax tracking output file}
\usage{
readCtrax(ctraxPath, sep = ",", mirrorY = FALSE, imgHeight = NA)
}
\arguments{
\item{ctraxPath}{The path of the Ctrax output file to load within R environment
(e.g. "C:/Users/\link{username}/Desktop/video_folder/Ctrax_output.mat")}

\item{mirrorY}{TRUE or FALSE, set the origin of Y coords, if TRUE Y coords are mirrored}

\item{imgHeight}{A numeric value expressed in pixels, the true length of Y axis
corresponding to the height of the image or video resolution (optional, only used when mirrorY = TRUE)}
}
\value{
A list containing tracking data
}
\description{
Given a .mat file corresponding to Ctrax tracking output,
this function open the .mat file in R, reconstruct time serie and if needed mirror Y coordinates
}
\examples{

# TODO

}
\seealso{
generate_time, mirrorY_coords
}
