% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrackR.R
\name{readTrackR}
\alias{readTrackR}
\title{Import TrackR tracking output .csv file.}
\usage{
readTrackR(trackRPath, flipY = FALSE, imgHeight = NULL, rawDat = FALSE)
}
\arguments{
\item{trackRPath}{The full path of the TrackR output file (.csv).}

\item{flipY}{A logical value (i.e., TRUE or FALSE) indicating whether the origin of y coordinates should be flipped. If TRUE, y coordinates are flipped to start on the top-left (default = FALSE).}

\item{imgHeight}{A numeric value expressed in pixels, the length of Y axis
corresponding to the height of the image or video resolution (optional, only used when flipY = TRUE).}

\item{rawDat}{A logical value (i.e., TRUE or FALSE) indicating whether all other elements retrieved from the tracking output should appended to the tracklets data (see \href{https://swarm-lab.github.io/trackR}{trackR}).
Note that this may drastically increase the size of the object returned by the function (default = FALSE).}
}
\value{
An object of class "tracklets" containing a list of tracklets and their characteristics classically used for further computations.
In case \link{rawDat} argument is TRUE, it also append all the others elements returned by the tracking software (for TrackR it corresponds to "n" the number of pixels covered by the object in a given frame).
Also, by default the function returns y coordinates starting on the bottom-left.
}
\description{
Given the path of the .csv file corresponding to TrackR output,
this function returns an object of class "tracklets", a list of tracklets (data frame) containing 7 elements classically used for further computations using MoveR package:
\itemize{
\item{'maj.ax': }{the length of the major axis (i.e., the midline) for a particle over frame (i.e., length of the ellipse).}
\item{'angle': }{the particle's absolute angle in radians, orientation of the particle according to y-axis.}
\item{'min.ax': }{the length of the minor axis for a particle over frame (i.e., width of the ellipse).}
\item{'x.pos': }{x position of the particle's centroid.}
\item{'y.pos': }{y position of the particle's centroid.}
\item{'identity': }{the particle's identity given by the tracking software.}
\item{'frame': }{the video frame number at which the measurements has been made.}
}

The function can also append all the others elements returned by the tracking software (see \link{rawDat} argument)
Also, the function can flip y coordinates (see \link{flipY} argument).
}
\examples{
\dontrun{

# Download the first dataset from the sample data repository
Path2Data <- MoveR::DLsampleData(dataSet = 1, tracker = "TrackR")
Path2Data

# Import the data as an object of class "tracklets"
# also do not flip Y coordinates (start on the bottom-left)
Data <-
  MoveR::readTrackR(Path2Data[[1]])
  
str(Data)

} 
}
\references{
\href{https://swarm-lab.github.io/trackR}{trackR - Multi-object tracking with R}
}
\seealso{
\code{\link{readAnimalTA}} \code{\link{readCtrax}}, \code{\link{readTrex}}, \code{\link{readIdtracker}}, \code{\link{flipYCoords}}
}
\author{
Quentin PETITJEAN
}
