% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrex.R
\name{readTrex}
\alias{readTrex}
\title{Load Trex tracking output .npz files}
\usage{
readTrex(trexPath, mirrorY = FALSE, imgHeight = 1080)
}
\arguments{
\item{trexPath}{The path of the Trex outputs folder to load .npz files within R environment
(e.g. "C:/Users/\link{username}/Desktop/Trex_output")}

\item{mirrorY}{TRUE or FALSE, set the origin of Y coords, if TRUE Y coords are mirrored}

\item{imgHeight}{A numeric value expressed in pixels, the true length of Y axis
corresponding to the height of the image or video resolution (default = 1080)}
}
\value{
A set of two lists containing tracking data, the first list corresponding to raw tracking data
in usable format for this package, the second corresponding to raw data as returned by Trex
}
\description{
Given the path of a folder containing multiple Trex outputs (one .npz file per individual),
this function group them in a list containing 2 dataframes' lists.
The first list containing 8 elements is the processed output (i.e., the elements further used by functions of this package).
The second list of 33 elements return the classic set of variables returned by Trex.
If needed this function can also mirror Y coordinates
}
\examples{

# TODO

}
\seealso{
readCtrax, mirrorYFunc
}
