% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrex.R
\name{readTrex}
\alias{readTrex}
\title{Load Trex tracking output .npz files.}
\usage{
readTrex(trexPath, mirrorY = TRUE, imgHeight = NULL, rawDat = FALSE)
}
\arguments{
\item{trexPath}{The path of the Trex output folder where .npz files are stored.}

\item{mirrorY}{TRUE or FALSE, set the origin of y coordinates, if TRUE y coordinates are mirrored to start on the bottom-left (default = TRUE).}

\item{imgHeight}{A numeric value expressed in pixels, the length of Y axis
corresponding to the height of the image or video resolution (optional, only used when mirrorY = TRUE).}

\item{rawDat}{TRUE or FALSE, if TRUE add a second list containing all the elements retrieved from .npz files (see \href{https://trex.run}{trex.run}),
may drastically increase the size of the object returned by the function (default = FALSE).}
}
\value{
A list containing either a list of 9 elements classically used for further computations or a list containing 2 sublists, the first corresponding to the one previously mentioned
and the second containing all the elements retrieved from the .npz files (see rawDat argument). Also, by default the function mirror y coordinates to start on the bottom-left.
}
\description{
Given the path of a folder containing multiple Trex outputs (one .npz file per particle/individual),
this function returns a list of 9 vectors classically used for further computations using MoveR package:
\itemize{
\item{'maj.ax': }{the length of the major axis (i.e., the midline) for a particle over frame (i.e., length of the ellipse).}
\item{'angle': }{the particle's absolute angle in radians, orientation of the particle according to X-axis.}
\item{'min.ax': }{the length of the minor axis for a particle over frame (i.e., width of the ellipse), returns NA since it is not present in the TREX output.}
\item{'x.pos': }{x position of the particle's centroid.}
\item{'y.pos': }{y position of the particle's centroid.}
\item{'identity': }{the particle's identity given by the tracking software.}
\item{'frame': }{the video frame number at which the measurements has been made.}
\item{'ntargets': }{the number of particle tracked over each frame.}
\item{'timestamps': }{the elapsed time over each frame, in seconds.}
}

Alternatively, the function can returns a list containing 2 sublists, the first corresponding to the one mentioned above
and the second containing all the elements retrieved from the .npz files (see rawDat argument).
Also, by default the function mirror y coordinates to start on the bottom-left (see mirrorY argument).
}
\examples{

# Load the list containing the 9 vectors classically used for further computation
# and mirror Y coordinates to start on the bottom-left

Data <-
  readTrex(
    system.file("sampleData/sample_1/TREXOutput", package = "MoveR"),
    mirrorY = T,
    imgHeight = 2160,
    rawDat = F
  )

# Load the list containing 2 sublists, the first containing the 9 vectors classically used for further computation
# and the second list containing all the elements retrieved from .npz files,
# also do not mirror Y coordinates (start on the top-left)

Data <-
  readTrex(
    system.file("sampleData/sample_1/TREXOutput", package = "MoveR"),
    mirrorY = F,
    imgHeight = NULL,
    rawDat = T
  )

}
\references{
Tristan Walter, Iain D Couzin (2021) TRex, a fast multi-animal tracking system with markerless identification, and 2D estimation of posture and visual fields eLife 10:e64000.
\href{https://trex.run}{trex.run}
}
\seealso{
\code{\link{readCtrax}}, \code{\link{readTrackR}}, \code{\link{readIdtracker}}, \code{\link{mirrorYFunc}}
}
\author{
Quentin PETITJEAN
}
