% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleFrags.R
\name{resampleFrags}
\alias{resampleFrags}
\title{Fragment resampling}
\usage{
resampleFrags(trackDat, Tstep)
}
\arguments{
\item{trackDat}{A list of data frame containing tracking informations for each fragment
(e.g., maj.ax, angle, min.ax, x.pos, y.pos, ...)}

\item{TStep}{A numeric value corresponding to the length of the resampling step (number of frames)}
}
\value{
A list of data frame containing resampled tracking informations for each fragment
}
\description{
Given a list of data frame containing tracking informations for each fragment,
this function returns a list of resampled fragments according to the length of the resampling
step specified by the user
}
\examples{

#TODO

}
