% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinuosity.R
\name{sinuosity}
\alias{sinuosity}
\title{Compute corrected rediscretized sinuosity index}
\usage{
sinuosity(df, scale = NULL, unit = NULL, segL)
}
\arguments{
\item{df}{A data frame containing x, y coordinates and time columns named "x.pos", "y.pos", "frame" for a fragment}

\item{scale}{A ratio corresponding to the scaling factor to be applied to the trajectory coordinates
(e.g., size in cm / size in pixels; see trajr::TrajScale())}

\item{unit}{The unit expected after scaling (e.g., "cm", "m", ...)}

\item{segL}{A numeric value expressed in the unit specified by user and corresponding
to the length of the resampling step needed to discretize the input trajectory
see trajr::TrajRediscretize())}
}
\value{
this function returns a vector containing the value of sinuosity for a given fragment
}
\description{
Given a data frames containing tracking informations for a given fragment,
this function rediscretize fragment path and returns a vector containing the value of
sinuosity along this fragment
}
\examples{

#TODO

}
