% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed.R
\name{speed}
\alias{speed}
\title{Compute speed}
\usage{
speed(df, scale = NULL, unit = NULL)
}
\arguments{
\item{df}{A data frame containing at x, y coordinates and time columns named "x.pos", "y.pos", "frame" for a fragment}

\item{scale}{A ratio corresponding to the scaling factor to be applied to the trajectory coordinates
(e.g., size in cm / size in pixels; see trajr::TrajScale())}

\item{unit}{The unit expected after scaling (e.g., "cm", "m", ...)}
}
\value{
this function returns a vector containing the value of speed along a given fragment
}
\description{
Given a data frames containing tracking informations for a given fragment,
this function scale fragment path and returns a vector containing the value of
speed along this fragment
}
\examples{

#TODO

}
