% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackStats.R
\name{trackStats}
\alias{trackStats}
\title{Summary statistics of the tracking data.}
\usage{
trackStats(
  trackDat,
  frameR = NULL,
  scale = NULL,
  units = NULL,
  progress = TRUE
)
}
\arguments{
\item{trackDat}{A list of data frame containing tracking information for each tracklet (i.e., x.pos, y.pos, frame).}

\item{frameR}{A numeric value expressed in frames per second, the frequency at which frames are recorded/displayed in the video
(optional).}

\item{scale}{A ratio corresponding to the scaling factor to be applied to the trajectory coordinates
(e.g., size in cm / size in pixels, (default = 1).}

\item{units}{A character string specifying the spatial unit of the coordinates after scaling (default = "pixels").}

\item{progress}{A logical value (i.e., TRUE or FALSE) indicating whether a progress bar should be displayed to inform process progression.}
}
\value{
A summary of video and tracklets data.
}
\description{
Given a list of tracking tracklets containing cartesian coordinates,
this function returns 2 sublists containing a summary of video and tracklets data:

\itemize{
\item{"VideoSummary": }{a list containing 4 elements:
\itemize{
\item{"videoDuration_f": }{the length of the video expressed in frames.}
\item{"videoDuration_s": }{the length of the video expressed in seconds.}
\item{"frameR": }{the frame rate of the video as specified by the user.}
\item{"scale": }{the scaling factor applied to the trajectory coordinates as specifed by the user.}
}
}

\item{"TrackletSummary": }{a list containing 8 elements:
\itemize{
\item{"trackNb": }{the total number of tracklets over the video (i.e., trajectories).}
\item{"totTrackDuration_f": }{the sum of the duration of all tracklets expressed in frames.}
\item{"totTrackDuration_s": }{the sum of the duration of all tracklets expressed in seconds}
\item{"totTrackLength_unit": }{the sum of the length of all tracklets in spatial unit as specified by the user (e.g., pixels, cm).}
\item{"trackId": }{the tracklets identity.}
\item{"trackDuration_f": }{the duration of each tracklets expressed in frames.}
\item{"trackDuration_s": }{the duration of each tracklets expressed in seconds.}
\item{"trackLength_unit": }{the length of each tracklets in spatial unit as specified by the user (e.g., pixels, cm).}
}
}
}
}
\examples{

set.seed(2023)
# generate some dummy tracklets
## start to specify some parameters to generate tracklets
TrackN <- 500 # the number of tracklet to simulate
TrackL <- 1:1000 # the length of the tracklets or a sequence to randomly sample tracklet length

TrackList <- stats::setNames(lapply(lapply(seq(TrackN), function(i)
  trajr::TrajGenerate(sample(TrackL, 1), random = TRUE, fps = 1)), function(j)
    data.frame(
      x.pos = j$x - min(j$x),
      y.pos = j$y - min(j$y),
      frame = j$time
    )), seq(TrackN))

# compute and display tracking summary
TrackSummary <- MoveR::trackStats(TrackList,
                                  frameR = 25,
                                  scale = 1,
                                  units = "pixels")

# retrieve the TrackLength and duration in frame from the summary "manually"
TrackL <- TrackSummary[["TrackletSummary"]][["trackLength_pixels"]]
TrackD <- TrackSummary[["TrackletSummary"]][["trackDuration_f"]]

# or using listGet utility
TrackL2 <- listGet(TrackSummary, "trackLength_pixels")
TrackD2 <- listGet(TrackSummary, "trackDuration_f")

# plot the distribution of tracklets length and duration
par(mfrow = c(2, 2))
hist(TrackL)
hist(TrackD)
hist(TrackL2)
hist(TrackD2)

}
\seealso{
\code{\link[trajr]{TrajScale}}, \code{\link[trajr]{TrajFromCoords}}
}
\author{
Quentin PETITJEAN
}
