% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnAngle.R
\name{turnAngle}
\alias{turnAngle}
\title{Compute turning angle}
\usage{
turnAngle(df, unit = c("radians", "degrees"))
}
\arguments{
\item{df}{A data frame containing x, y coordinates columns named "x.pos", "y.pos" for a given fragment}

\item{unit}{A character string indicating whether the function should returns turning angle in radians or degrees}
}
\value{
this function returns a vector containing the value of turning angle computed along a given fragment
}
\description{
Given a data frames containing tracking information for a particle,
this function compute returns a vector containing the value of turning angle along the considered fragment
}
\examples{

#TODO

}
